requirements.txt
apparmor==3.0.1
appdirs==1.4.4
application-utility==1.3.2
asgiref==3.3.1
asn1crypto==1.4.0
attrs==20.3.0
autopep8==1.5.4
backcall==0.2.0
bcrypt==3.2.0
btrfsutil==5.10
CacheControl==0.12.6
cached-property==1.5.2
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==3.3.1
cupshelpers==1.0
decorator==4.4.2
distlib==0.3.1
distro==1.5.0
dj-database-url==0.5.0
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
docker==4.4.1
docker-compose==1.28.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
filelock==3.0.12
html5lib==1.1
idna==2.10
importlib-metadata==3.4.0
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
jsonschema==3.2.0
kazam==1.4.5
keyutils==0.6
LibAppArmor==3.0.1
louis==3.16.0
msgpack==1.0.2
netsnmp-python==1.0a1
npyscreen==4.10.5
ordered-set==4.0.2
packaging==20.8
pacman-mirrors==4.19.1
paramiko==2.7.2
parso==0.8.1
pep517==0.9.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
ply==3.11
progress==1.5
prompt-toolkit==3.0.14
ptyprocess==0.7.0
pwquality==1.4.4
pycairo==1.20.0
pycodestyle==2.6.0
pycparser==2.20
pycups==2.0.1
pycurl==7.43.0.6
Pygments==2.7.4
PyGObject==3.38.0
PyJWT==1.7.1
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyrsistent==0.17.3
pysmbc==1.0.23
python-decouple==3.4
python-distutils-extra==2.39
python-dotenv==0.15.0
python-libtorrent==1.2.11
pytz==2020.5
pyxdg==0.26
PyYAML==5.3.1
rados==2.0.0
rbd==2.0.0
reportlab==3.5.60
requests==2.25.1
resolvelib==0.5.4
retrying==1.3.3
rgw==2.0.0
sip==4.19.24
six==1.15.0
sqlparse==0.4.1
team==1.0
texttable==1.6.3
toml==0.10.2
traitlets==5.0.5
udiskie==2.3.2
urllib3==1.26.3
virtualenv==20.4.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
zipp==3.4.0
gunicorn
psycopg2
